move Client{} type from mockserver test to it's own directory
